% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Isoviewer App Server}
\usage{
app_server(data_dir)
}
\arguments{
\item{data_dir}{the directory for local data files}
}
\description{
Generates the server part of the isoviewer app
}
