% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Isoviewer App UI}
\usage{
app_ui(sidebar_width = 150, default_menu = "welcome")
}
\arguments{
\item{sidebar_width}{the width of the sidebar}

\item{deault_menu}{default selected menu}
}
\description{
Generates the user interface part of the isoviewer app
}
