% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cf_raw_data.R
\name{cfRawDataPlotUI}
\alias{cfRawDataPlotUI}
\title{Continuous Flow Raw Data Plot UI}
\usage{
cfRawDataPlotUI(id)
}
\arguments{
\item{id}{the module id}
}
\description{
Continuous Flow Raw Data Plot UI
}
\seealso{
Other continuous flow raw data module functions: \code{\link{cfRawDataSelectorUI}},
  \code{\link{cfRawDataServer}},
  \code{\link{cfRawDataSettingsUI}}
}
