% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cf_raw_data.R
\name{cfRawDataSelectorUI}
\alias{cfRawDataSelectorUI}
\title{Continuous Flow Raw Data Selector UI}
\usage{
cfRawDataSelectorUI(id, width = 4, selector_height = "200px")
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Continuous Flow Raw Data Selector UI
}
\seealso{
Other continuous flow raw data module functions: \code{\link{cfRawDataPlotUI}},
  \code{\link{cfRawDataServer}},
  \code{\link{cfRawDataSettingsUI}}
}
