% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cf_raw_data.R
\name{cfRawDataServer}
\alias{cfRawDataServer}
\title{Continuous Flow Raw Data}
\usage{
cfRawDataServer(input, output, session, isofiles, dataset_name,
  visible = NULL)
}
\arguments{
\item{isofiles}{reactive function returning the currently loaded isofiles}

\item{visible}{reactive function determining visibility of the auxiliary boxes}
}
\description{
Continuous Flow Raw Data
}
\seealso{
Other continuous flow raw data module functions: \code{\link{cfRawDataPlotUI}},
  \code{\link{cfRawDataSelectorUI}},
  \code{\link{cfRawDataSettingsUI}}
}
