% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cf_raw_data.R
\name{cfRawDataSettingsUI}
\alias{cfRawDataSettingsUI}
\title{Continuous Flow Raw Data Settings UI}
\usage{
cfRawDataSettingsUI(id, width = 4)
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Continuous Flow Raw Data Settings UI
}
\seealso{
Other continuous flow raw data module functions: \code{\link{cfRawDataPlotUI}},
  \code{\link{cfRawDataSelectorUI}},
  \code{\link{cfRawDataServer}}
}
