% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_code_preview.R
\name{code_preview_server}
\alias{code_preview_server}
\title{Code Preview Server}
\usage{
code_preview_server(input, output, session, code_func_reac, download_file)
}
\arguments{
\item{code_func_reac}{reactive function that returns a function(!) with parameters \code{rmarkdown} and \code{front_matter} that can be called to generate the code}

\item{download_file}{reactive function that returns the download file name (with or without .Rmd ending)}
}
\description{
Code Preview Server
}
\seealso{
Other code preview module functions: 
\code{\link{code_preview_shinyjs_extension}()},
\code{\link{code_preview_ui}()}
}
\concept{code preview module functions}
