% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_code_preview.R
\name{code_preview_shinyjs_extension}
\alias{code_preview_shinyjs_extension}
\title{JS Extension for code preview}
\usage{
code_preview_shinyjs_extension()
}
\description{
Generates the shinyJS extensions for updating the code preview. Must be included ONCE in the UI if this module is used (no matter how many times).
}
\seealso{
Other code preview module functions: 
\code{\link{code_preview_server}()},
\code{\link{code_preview_ui}()}
}
\concept{code preview module functions}
