% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_code_preview.R
\name{code_preview_ui}
\alias{code_preview_ui}
\title{Code Preview UI}
\usage{
code_preview_ui(id, width = 12, height = "400px")
}
\arguments{
\item{id}{the module id}

\item{width}{width of the box}

\item{height}{height of the code preview}
}
\description{
Code Preview UI
}
\seealso{
Other code preview module functions: 
\code{\link{code_preview_server}()},
\code{\link{code_preview_shinyjs_extension}()}
}
\concept{code preview module functions}
