% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_code_preview.R
\name{codePreviewShinyjsExtension}
\alias{codePreviewShinyjsExtension}
\title{JS Extension for code preview}
\usage{
codePreviewShinyjsExtension()
}
\description{
Generates the shinyJS extensions for updating the code preview. Must be included ONCE in the UI if this module is used (no matter how many times).
}
\seealso{
Other code preview module functions: \code{\link{codePreviewServer}},
  \code{\link{codePreviewUI}}
}
