% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_code_preview.R
\name{codePreviewUI}
\alias{codePreviewUI}
\title{Code Preview UI}
\usage{
codePreviewUI(id, width = 12, height = "400px")
}
\arguments{
\item{id}{the module id}

\item{width}{width of the box}

\item{height}{height of the code preview}
}
\description{
Code Preview UI
}
\seealso{
Other code preview module functions: \code{\link{codePreviewServer}},
  \code{\link{codePreviewShinyjsExtension}}
}
