% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cf_data.R
\name{continuousFlowDataServer}
\alias{continuousFlowDataServer}
\title{Continuous Flow Data Server}
\usage{
continuousFlowDataServer(input, output, session, isofiles, dataset_name)
}
\arguments{
\item{isofiles}{reactive function returning the currently loaded isofiles}

\item{dataset_name}{reactive function with the dataset name}
}
\description{
Continuous Flow Data Server
}
\seealso{
Other view continuous flow module functions: \code{\link{continuousFlowDataUI}},
  \code{\link{continuousFlowViewServer}},
  \code{\link{continuousFlowViewUI}}
}
