% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cf_data.R
\name{continuousFlowDataUI}
\alias{continuousFlowDataUI}
\title{Continuous Flow Data UI}
\usage{
continuousFlowDataUI(id, width = 12)
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Continuous Flow Data UI
}
\seealso{
Other view continuous flow module functions: \code{\link{continuousFlowDataServer}},
  \code{\link{continuousFlowViewServer}},
  \code{\link{continuousFlowViewUI}}
}
