% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cf_view.R
\name{continuousFlowViewServer}
\alias{continuousFlowViewServer}
\title{Continuous Flow View Server}
\usage{
continuousFlowViewServer(input, output, session, data_dir, load_server = NULL)
}
\arguments{
\item{data_dir}{the directory for local data files}

\item{load_server}{if a load server is present}

\item{isofiles}{reactive function returning the currently loaded isofiles}
}
\description{
Continuous Flow View Server
}
\seealso{
Other view continuous flow module functions: \code{\link{continuousFlowDataServer}},
  \code{\link{continuousFlowDataUI}},
  \code{\link{continuousFlowViewUI}}
}
