% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cf_view.R
\name{continuousFlowViewUI}
\alias{continuousFlowViewUI}
\title{Continuous Flow View UI}
\usage{
continuousFlowViewUI(id, width = 12)
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Continuous Flow View UI
}
\seealso{
Other view continuous flow module functions: \code{\link{continuousFlowDataServer}},
  \code{\link{continuousFlowDataUI}},
  \code{\link{continuousFlowViewServer}}
}
