% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_download.R
\name{data_download_server}
\alias{data_download_server}
\title{data download server}
\usage{
data_download_server(input, output, session, settings, get_variable)
}
\description{
data download server
}
