% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table.R
\name{data_table_column_selector_ui}
\alias{data_table_column_selector_ui}
\title{Column Selector UI}
\usage{
data_table_column_selector_ui(
  id,
  width = 4,
  pre_table = list(),
  post_table = list()
)
}
\arguments{
\item{width}{box width}
}
\description{
Column Selector UI
}
