% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table_file_info.R
\name{data_table_file_info_ui}
\alias{data_table_file_info_ui}
\title{File Info Table UI}
\usage{
data_table_file_info_ui(id, ...)
}
\arguments{
\item{...}{passed on to data_table_ui}
}
\description{
File Info Table UI
}
