% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table_raw_data.R
\name{data_table_raw_data_column_selector_ui}
\alias{data_table_raw_data_column_selector_ui}
\title{Raw Data Column Seletor UI}
\usage{
data_table_raw_data_column_selector_ui(id, ...)
}
\arguments{
\item{...}{passed on to data_table_column_selector_ui}
}
\description{
Raw Data Column Seletor UI
}
