% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table_raw_data.R
\name{data_table_raw_data_ui}
\alias{data_table_raw_data_ui}
\title{Raw Data Table UI}
\usage{
data_table_raw_data_ui(id, ...)
}
\arguments{
\item{...}{passed on to data_table_ui}
}
\description{
Raw Data Table UI
}
