% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table_resistors.R
\name{data_table_resistors_server}
\alias{data_table_resistors_server}
\title{Resistors Server}
\usage{
data_table_resistors_server(
  input,
  output,
  session,
  get_variable,
  get_iso_files,
  is_visible
)
}
\arguments{
\item{get_variable}{reactive function returning the selected variable}

\item{get_iso_files}{reactive function returning the currently loaded isofiles}

\item{is_visible}{reactive function determining visibility of the auxiliary boxes}
}
\description{
Resistors Server
}
