% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table_resistors.R
\name{data_table_resistors_ui}
\alias{data_table_resistors_ui}
\title{Resistors Table UI}
\usage{
data_table_resistors_ui(id, ...)
}
\arguments{
\item{...}{passed on to data_table_ui}
}
\description{
Resistors Table UI
}
