% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table.R
\name{data_table_server}
\alias{data_table_server}
\title{Basic data table server with column selection option}
\usage{
data_table_server(
  input,
  output,
  session,
  settings,
  get_variable,
  get_iso_files,
  is_visible,
  get_data_table,
  get_data_table_columns
)
}
\arguments{
\item{get_variable}{reactive function returning the selected variable}

\item{get_iso_files}{reactive function returning the currently loaded isofiles}

\item{is_visible}{reactive function determining visibility of the auxiliary boxes}

\item{get_data_table}{a regular or reactive function taking iso_files and a list of column names to retrieve the data table}

\item{get_data_table_columns}{a regular function taking iso_files and returning a vector of data table columns}
}
\description{
This is for the usual iso_get_... aggregation functions.
}
\concept{file info module functions}
