% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table_standards.R
\name{data_table_standards_column_selector_ui}
\alias{data_table_standards_column_selector_ui}
\title{Standards Column Seletor UI}
\usage{
data_table_standards_column_selector_ui(id, ...)
}
\arguments{
\item{...}{passed on to data_table_column_selector_ui}
}
\description{
Standards Column Seletor UI
}
