% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table_standards.R
\name{data_table_standards_ui}
\alias{data_table_standards_ui}
\title{Standards Table UI}
\usage{
data_table_standards_ui(id, ...)
}
\arguments{
\item{...}{passed on to data_table_ui}
}
\description{
Standards Table UI
}
