% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table_vendor_data_table.R
\name{data_table_vendor_data_table_ui}
\alias{data_table_vendor_data_table_ui}
\title{Vendor Data  Table UI}
\usage{
data_table_vendor_data_table_ui(id, ...)
}
\arguments{
\item{...}{passed on to data_table_ui}
}
\description{
Vendor Data  Table UI
}
