% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_datasets.R
\name{datasetsServer}
\alias{datasetsServer}
\title{Datasets server}
\usage{
datasetsServer(input, output, session, data_dir, extensions, load_func,
  saved_datasets)
}
\arguments{
\item{data_dir}{the directory for local data files}

\item{extensions}{which extensions to allow}

\item{load_func}{the loading function (as character!)}

\item{saved_datasets}{reactive function returning a vector of saved datasets (with names as the relative path)}
}
\description{
Datasets server
}
\seealso{
Other datasets module functions: \code{\link{datasetsUI}},
  \code{\link{problemsButton}},
  \code{\link{problemsServer}}
}
