% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_datasets.R
\name{datasetsUI}
\alias{datasetsUI}
\title{Dataset UI}
\usage{
datasetsUI(id, width = 12, file_list_height = "200px")
}
\arguments{
\item{id}{the module id}

\item{width}{box width}

\item{file_list_height}{height of the file list}
}
\description{
Dataset UI
}
\seealso{
Other datasets module functions: \code{\link{datasetsServer}},
  \code{\link{problemsButton}},
  \code{\link{problemsServer}}
}
