% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table.R
\name{dataTableServer}
\alias{dataTableServer}
\title{Data Table Server}
\usage{
dataTableServer(input, output, session, rownames = FALSE,
  selection = "none", filter = "top", sig_digits = 3,
  extensions = c("Buttons", "KeyTable"), options = list(dom = "Bltpi",
  pageLength = 5, lengthMenu = c(5, 10, 15, 20, 50, 100), keys = TRUE, scrollX =
  TRUE, buttons = list(list(extend = "colvis"))), ...)
}
\arguments{
\item{rownames}{default passed onto datatable function}

\item{selection}{default passed onto datatable function}

\item{filter}{default passed onto datatable function}

\item{extensions}{default passed onto datatable function}

\item{options}{default passed onto datatable function}

\item{...}{additional params passed onto datatable function}

\item{sig_digitis}{number of significant digits used for all numeric formatting}
}
\description{
Data Table Server
}
