% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table.R
\name{dataTableServerSimple}
\alias{dataTableServerSimple}
\title{Simple data table with useful defaults for a simple table}
\usage{
dataTableServerSimple(input, output, session, rownames = FALSE,
  selection = "none", filter = "none", sig_digits = 3,
  extensions = list(), options = list(dom = "t", pageLength = 50), ...)
}
\arguments{
\item{rownames}{default passed onto datatable function}

\item{selection}{default passed onto datatable function}

\item{filter}{default passed onto datatable function}

\item{extensions}{default passed onto datatable function}

\item{options}{default passed onto datatable function}

\item{...}{additional params passed onto datatable function}
}
\description{
Simple data table with useful defaults for a simple table
}
