% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_upload.R
\name{dataUploadUI}
\alias{dataUploadUI}
\title{Data Upload UI}
\usage{
dataUploadUI(id, open_label = "Upload data",
  open_tooltip = "Upload local data files to the server",
  dialog_text = NULL, accept = c("application/zip", ".zip"))
}
\arguments{
\item{id}{data upload id}

\item{open_label}{the label for the link to open the modal dialog}

\item{dialog_text}{what should the dialog say}

\item{accept}{which mime types to allow (e.g. zip is c('application/zip', '.zip'))}
}
\description{
Data Upload UI
}
\seealso{
Other data upload module functions: \code{\link{dataUploadServer}}
}
