% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_di_raw_data.R
\name{diRawDataPlotUI}
\alias{diRawDataPlotUI}
\title{Dual Inlet Raw Data Plot UI}
\usage{
diRawDataPlotUI(id)
}
\arguments{
\item{id}{the module id}
}
\description{
Dual Inlet Raw Data Plot UI
}
\seealso{
Other dual inlet raw data module functions: \code{\link{diRawDataSelectorUI}},
  \code{\link{diRawDataServer}},
  \code{\link{diRawDataSettingsUI}}
}
