% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_di_raw_data.R
\name{diRawDataSelectorUI}
\alias{diRawDataSelectorUI}
\title{Dual Inlet Raw Data Selector UI}
\usage{
diRawDataSelectorUI(id, width = 4, selector_height = "200px")
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Dual Inlet Raw Data Selector UI
}
\seealso{
Other dual inlet raw data module functions: \code{\link{diRawDataPlotUI}},
  \code{\link{diRawDataServer}},
  \code{\link{diRawDataSettingsUI}}
}
