% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_di_raw_data.R
\name{diRawDataServer}
\alias{diRawDataServer}
\title{Dual Inlet Raw Data}
\usage{
diRawDataServer(input, output, session, isofiles, dataset_name,
  visible = NULL)
}
\arguments{
\item{isofiles}{reactive function returning the currently loaded isofiles}

\item{visible}{reactive function determining visibility of the auxiliary boxes}
}
\description{
Dual Inlet Raw Data
}
\seealso{
Other dual inlet raw data module functions: \code{\link{diRawDataPlotUI}},
  \code{\link{diRawDataSelectorUI}},
  \code{\link{diRawDataSettingsUI}}
}
