% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_di_raw_data.R
\name{diRawDataSettingsUI}
\alias{diRawDataSettingsUI}
\title{Dual Inlet Raw Data Settings UI}
\usage{
diRawDataSettingsUI(id, width = 4)
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Dual Inlet Raw Data Settings UI
}
\seealso{
Other dual inlet raw data module functions: \code{\link{diRawDataPlotUI}},
  \code{\link{diRawDataSelectorUI}},
  \code{\link{diRawDataServer}}
}
