% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_di_data.R
\name{dualInletDataServer}
\alias{dualInletDataServer}
\title{Dual Inlet Data Server}
\usage{
dualInletDataServer(input, output, session, isofiles)
}
\arguments{
\item{isofiles}{reactive function returning the currently loaded isofiles}
}
\description{
Dual Inlet Data Server
}
\seealso{
Other view dual inlet module functions: \code{\link{dualInletDataUI}},
  \code{\link{dualInletViewServer}},
  \code{\link{dualInletViewUI}}
}
