% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_di_data.R
\name{dualInletDataUI}
\alias{dualInletDataUI}
\title{Dual Inlet Data UI}
\usage{
dualInletDataUI(id, width = 12)
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Dual Inlet Data UI
}
\seealso{
Other view dual inlet module functions: \code{\link{dualInletDataServer}},
  \code{\link{dualInletViewServer}},
  \code{\link{dualInletViewUI}}
}
