% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_di_view.R
\name{dualInletViewServer}
\alias{dualInletViewServer}
\title{Dual Inlet View Server}
\usage{
dualInletViewServer(input, output, session, data_dir, load_server = NULL)
}
\arguments{
\item{data_dir}{the directory for local data files}

\item{load_server}{if a load server is present}

\item{isofiles}{reactive function returning the currently loaded isofiles}
}
\description{
Dual Inlet View Server
}
\seealso{
Other view dual inlet module functions: \code{\link{dualInletDataServer}},
  \code{\link{dualInletDataUI}},
  \code{\link{dualInletViewUI}}
}
