% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_di_view.R
\name{dualInletViewUI}
\alias{dualInletViewUI}
\title{Dual Inlet View UI}
\usage{
dualInletViewUI(id, width = 12)
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Dual Inlet View UI
}
\seealso{
Other view dual inlet module functions: \code{\link{dualInletDataServer}},
  \code{\link{dualInletDataUI}},
  \code{\link{dualInletViewServer}}
}
