% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_export.R
\name{exportServer}
\alias{exportServer}
\title{Export Server}
\usage{
exportServer(input, output, session, isofiles, dataset_name, visible = NULL)
}
\arguments{
\item{isofiles}{reactive function returning the currently loaded isofiles}

\item{visible}{reactive function determining visibility of the auxiliary boxes}
}
\description{
Export Server
}
\seealso{
Other export module: \code{\link{exportSettingsUI}},
  \code{\link{exportUI}}
}
