% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_export.R
\name{exportSettingsUI}
\alias{exportSettingsUI}
\title{Export Setings UI}
\usage{
exportSettingsUI(id, width = 4)
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Export Setings UI
}
\seealso{
Other export module: \code{\link{exportServer}},
  \code{\link{exportUI}}
}
