% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_export.R
\name{exportUI}
\alias{exportUI}
\title{Method Info Table UI}
\usage{
exportUI(id)
}
\arguments{
\item{id}{the module id}
}
\description{
Method Info Table UI
}
\seealso{
Other export module: \code{\link{exportServer}},
  \code{\link{exportSettingsUI}}
}
