% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_info.R
\name{fileInfoSelectorUI}
\alias{fileInfoSelectorUI}
\title{File Info Selector UI}
\usage{
fileInfoSelectorUI(id, width = 4, selector_height = "200px")
}
\arguments{
\item{id}{the module id}

\item{width}{box width}

\item{selector_height}{file selector height}
}
\description{
File Info Selector UI
}
\seealso{
Other file info module functions: \code{\link{fileInfoServer}},
  \code{\link{fileInfoTableUI}},
  \code{\link{methodInfoSelectorUI}}
}
