% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_info.R
\name{fileInfoServer}
\alias{fileInfoServer}
\title{File Info Server}
\usage{
fileInfoServer(input, output, session, isofiles, visible = NULL)
}
\arguments{
\item{isofiles}{reactive function returning the currently loaded isofiles}

\item{visible}{reactive function determining visibility of the auxiliary boxes}
}
\description{
File Info Server
}
\seealso{
Other file info module functions: \code{\link{fileInfoSelectorUI}},
  \code{\link{fileInfoTableUI}},
  \code{\link{methodInfoSelectorUI}}
}
