% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_info.R
\name{fileInfoTableUI}
\alias{fileInfoTableUI}
\title{File Info Table UI}
\usage{
fileInfoTableUI(id)
}
\arguments{
\item{id}{the module id}
}
\description{
File Info Table UI
}
\seealso{
Other file info module functions: \code{\link{fileInfoSelectorUI}},
  \code{\link{fileInfoServer}},
  \code{\link{methodInfoSelectorUI}}
}
