% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_browser.R
\name{fileSelectorServer}
\alias{fileSelectorServer}
\title{file selector server side}
\usage{
fileSelectorServer(input, output, session, root, root_name = basename(root),
  pattern = NULL, multiple = TRUE, enable_recent = TRUE,
  start_recent = TRUE, start_n_recent = 50, exclude_recent = NULL,
  start_sort_desc = TRUE)
}
\arguments{
\item{root}{directory}

\item{root_name}{the root directory name}

\item{pattern}{regexp file selection pattern}

\item{multiple}{whether to allow multiple file selection}

\item{enable_recent}{whether to have a recent tab}

\item{start_recent}{whether to start on the recent tab}

\item{start_n_recent}{how many recent items to show at the beginning}

\item{exclude_recent}{regex what to explude in the recent list (by default nothing is excluded)}

\item{start_sort_desc}{whether to start sorting in descending order}
}
\description{
file selector server side
}
\seealso{
Other file selection module functions: \code{\link{fileSelectorUI}},
  \code{\link{modalFileSelectorUI}}
}
