% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_browser.R
\name{fileSelectorUI}
\alias{fileSelectorUI}
\title{File Selector UI}
\usage{
fileSelectorUI(id, size = 12)
}
\arguments{
\item{id}{the file selector id}

\item{size}{the height of the selection box (number of rows)}
}
\description{
File Selector UI
}
\seealso{
Other file selection module functions: \code{\link{fileSelectorServer}},
  \code{\link{modalFileSelectorUI}}
}
