% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_function_plot_param.R
\name{function_plot_param_ui}
\alias{function_plot_param_ui}
\title{function plot parameter UI}
\usage{
function_plot_param_ui(id, label, label_width = 4, value_width = 8)
}
\description{
function plot parameter UI
}
