% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{gdrive_storage}
\alias{gdrive_storage}
\title{Create Google Drive remote storage}
\usage{
gdrive_storage(root, root_name = "Data", oauth_token = "gdrive_token.rds",
  local_dir = tempdir())
}
\arguments{
\item{root}{the google drive path that will serve as the root directory for data storage}

\item{oauth_token}{the file name for the authentication token (will be stored in the cache folder)}

\item{local_dir}{the local directory where temporary files from the remote should be stored}
}
\description{
This creates a google drive storage object as a remote data directory for the isoviewer GUI.
Only files and folders under the \code{root} directory will be accessible by isoviewer.
}
\seealso{
Other data storage functions: \code{\link{get_directory_tree}},
  \code{\link{get_recent_files}},
  \code{\link{init_storage}}, \code{\link{local_storage}},
  \code{\link{refresh_directory_tree}},
  \code{\link{refresh_recent_files}}
}
