% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{get_directory_tree}
\alias{get_directory_tree}
\title{Get the directory tree
Get directory tree from the storage object}
\usage{
get_directory_tree(storage)
}
\description{
Get the directory tree
Get directory tree from the storage object
}
\seealso{
Other data storage functions: \code{\link{gdrive_storage}},
  \code{\link{get_recent_files}},
  \code{\link{init_storage}}, \code{\link{local_storage}},
  \code{\link{refresh_directory_tree}},
  \code{\link{refresh_recent_files}}
}
