% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{get_recent_files}
\alias{get_recent_files}
\title{Get the recent files
Get recent files from the storage object}
\usage{
get_recent_files(storage)
}
\description{
Get the recent files
Get recent files from the storage object
}
\seealso{
Other data storage functions: \code{\link{gdrive_storage}},
  \code{\link{get_directory_tree}},
  \code{\link{init_storage}}, \code{\link{local_storage}},
  \code{\link{refresh_directory_tree}},
  \code{\link{refresh_recent_files}}
}
