% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{init_storage}
\alias{init_storage}
\title{Initialize storage}
\usage{
init_storage(storage, save_credentials = TRUE)
}
\arguments{
\item{storage}{the storage object}

\item{save_credentials}{if login credentials are required - should they be stored for next time?}
}
\description{
Initializes a storage object and checks that all necessary information for it to work properly is provided.
Throws errors if anything is amiss. Can be called explicitly (or implictly by calling any other operation to
work with the storage). Will only do the initializing once and afterwards just return the storage object.
}
\seealso{
Other data storage functions: \code{\link{gdrive_storage}},
  \code{\link{get_directory_tree}},
  \code{\link{get_recent_files}},
  \code{\link{local_storage}},
  \code{\link{refresh_directory_tree}},
  \code{\link{refresh_recent_files}}
}
