% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iso_find_objects}
\alias{iso_find_objects}
\title{Find iso objects}
\usage{
iso_find_objects(env = .GlobalEnv)
}
\arguments{
\item{env}{where to look for iso objects, by default the global environment}
}
\value{
list of found iso objects (all types) and their values
}
\description{
Finds all iso file objects in the provided environment (the global environment by default).
}
