% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{iso_start_viewer}
\alias{iso_start_viewer}
\title{Start the isoviewer graphical user interface}
\usage{
iso_start_viewer(
  iso_files = NULL,
  ...,
  restore = TRUE,
  reset = FALSE,
  log = FALSE,
  launch = TRUE
)
}
\arguments{
\item{iso_files}{variable name - if provided, will start the viewer on this collection of iso files}

\item{...}{passed on to the \code{\link[shiny]{runApp}} call (only if \code{launch = TRUE}), can include server-specific parameters such as host or port}

\item{restore}{whether to restore the GUI to the previous state (if available). If \code{TRUE}, will store the GUI settings for future restore.}

\item{reset}{whether to reset the GUI state upon viewer start}

\item{log}{whether to show log info messages or not}

\item{launch}{whether to launch the shiny app or return it as an object}
}
\description{
Searches for all iso file objects in the workspace and provides a graphical interface to exlore the data stored in them.
}
