% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_load_isofiles.R
\name{isofilesLoadServer}
\alias{isofilesLoadServer}
\title{Isofiles Load Server}
\usage{
isofilesLoadServer(input, output, session, data_dir, extensions, load_func,
  load_params = c(), allow_data_upload = FALSE, store_data = TRUE,
  collections_dir = file.path(data_dir, "collections"))
}
\arguments{
\item{data_dir}{the directory for local data files}

\item{extensions}{which extensions to allow}

\item{load_func}{the loading function (as character!)}

\item{load_params}{the loading checkboxes (parameter names and labels)}

\item{allow_data_upload}{whether to allow uploading of data}

\item{store_data}{whether data files (including .rda exports) are stored permanently (TRUE) or just temporarily (FALSE)}

\item{collections_dir}{the director for collection files}
}
\description{
Isofiles Load Server
}
\seealso{
Other isofiles load module functions: \code{\link{isofilesLoadUI}}
}
