% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_load_isofiles.R
\name{isofilesLoadUI}
\alias{isofilesLoadUI}
\title{Isofiles Load UI}
\usage{
isofilesLoadUI(id, label = NULL)
}
\arguments{
\item{id}{the module id}

\item{label}{what to label boxes with}
}
\description{
Isofiles Load UI
}
\seealso{
Other isofiles load module functions: \code{\link{isofilesLoadServer}}
}
