% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isoviewer-package}
\alias{isoviewer}
\alias{isoviewer-package}
\title{isoviewer: Visualize IRMS data}
\description{
Graphical user interface components for exploring isotope data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/isoverse/isoviewer}
  \item Report bugs at \url{https://github.com/isoverse/isoviewer/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
