% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{local_storage}
\alias{local_storage}
\title{Local file system storage}
\usage{
local_storage(root, root_name = "Data")
}
\arguments{
\item{root}{the local directory (relative to working directory or absolute) that will serve as the root directory for data storage}

\item{root_name}{the name of the root directory that should be shonw in isoviewer}
}
\description{
This creates a storage object for local data storage for the isoviewer GUI.
Only files and folders under hte \code{root} directory will be accessible by isoviewer.
}
\details{
The \code{\link{run}} function creates local storage automatically if only provided with a path.
}
\seealso{
Other data storage functions: \code{\link{gdrive_storage}},
  \code{\link{get_directory_tree}},
  \code{\link{get_recent_files}},
  \code{\link{init_storage}},
  \code{\link{refresh_directory_tree}},
  \code{\link{refresh_recent_files}}
}
