% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_method_info.R
\name{method_info_selector_ui}
\alias{method_info_selector_ui}
\title{Method Info Selector UIUI}
\usage{
method_info_selector_ui(id, width = 4)
}
\arguments{
\item{width}{box width}
}
\description{
Method Info Selector UIUI
}
\seealso{
Other file info module functions: 
\code{\link{data_table_server}()}
}
\concept{file info module functions}
