% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_method_info.R
\name{method_info_server}
\alias{method_info_server}
\title{Method Info Server}
\usage{
method_info_server(
  input,
  output,
  session,
  get_variable,
  get_iso_files,
  is_visible
)
}
\description{
Method Info Server
}
