% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_method_info.R
\name{method_info_table_ui}
\alias{method_info_table_ui}
\title{Method Info Table UI}
\usage{
method_info_table_ui(id, min_height = "800px;")
}
\description{
Method Info Table UI
}
\concept{method info module functions}
