% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_method_info.R
\name{methodInfoSelectorUI}
\alias{methodInfoSelectorUI}
\title{Method Info Selector UI}
\usage{
methodInfoSelectorUI(id, width = 4)
}
\arguments{
\item{id}{the module id}

\item{width}{box width}
}
\description{
Method Info Selector UI
}
\seealso{
Other file info module functions: \code{\link{fileInfoSelectorUI}},
  \code{\link{fileInfoServer}},
  \code{\link{fileInfoTableUI}}
}
