% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_method_info.R
\name{methodInfoServer}
\alias{methodInfoServer}
\title{Method Info Server}
\usage{
methodInfoServer(input, output, session, isofiles, visible = NULL)
}
\arguments{
\item{isofiles}{reactive function returning the currently loaded isofiles}

\item{visible}{reactive function determining visibility of the auxiliary boxes}
}
\description{
Method Info Server
}
\seealso{
Other method info module functions: \code{\link{methodInfoTableUI}}
}
