% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_method_info.R
\name{methodInfoTableUI}
\alias{methodInfoTableUI}
\title{Method Info Table UI}
\usage{
methodInfoTableUI(id)
}
\arguments{
\item{id}{the module id}
}
\description{
Method Info Table UI
}
\seealso{
Other method info module functions: \code{\link{methodInfoServer}}
}
