% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_browser.R
\name{modalFileSelectorUI}
\alias{modalFileSelectorUI}
\title{modalFileSelector}
\usage{
modalFileSelectorUI(id, size, dialog_size = "large",
  open_label = "Select file", dialog_label = open_label,
  close_label = "Select", link_wrapper = identity)
}
\arguments{
\item{id}{the file selector id}

\item{size}{the height of the selection box (number of rows)}

\item{dialog_size}{the size of the dialog, default is small}

\item{open_label}{the label for the link to open the modal dialog}

\item{dialog_label}{the label of the modal dialog, same as the open_label by default}

\item{close_label}{the label for the button to close the modal dialog}
}
\description{
modalFileSelector
}
