% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_di.R
\name{module_data_di_server}
\alias{module_data_di_server}
\title{Dual inlet flow files Server}
\usage{
module_data_di_server(
  input,
  output,
  session,
  settings,
  iso_objects = list(),
  get_selected_variable
)
}
\arguments{
\item{get_selected_variable}{reactive function returning the variable name}
}
\description{
Dual inlet flow files Server
}
