% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_scan.R
\name{module_data_scan_server}
\alias{module_data_scan_server}
\title{Scan flow files Server}
\usage{
module_data_scan_server(
  input,
  output,
  session,
  settings,
  iso_objects = list(),
  get_selected_variable
)
}
\arguments{
\item{get_selected_variable}{reactive function returning the variable name}
}
\description{
Scan flow files Server
}
