% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{module_data_server}
\alias{module_data_server}
\title{Data server (main module)}
\usage{
module_data_server(
  input,
  output,
  session,
  settings,
  iso_objects = list(),
  get_selected_variable,
  data_type = c("continuous_flow", "dual_inlet", "scan"),
  get_code_update = reactive({     function(rmarkdown) {         ""     } })
)
}
\arguments{
\item{get_selected_variable}{reactive function returning the variable name}

\item{data_type}{the type of data server (character), is used to construct iso_read... and iso_is_... calls}

\item{get_code_update}{a reactive code update function, has to return a function that takes 'rmarkdown' as a parameter}
}
\description{
Data server (main module)
}
