% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_selector.R
\name{module_file_selector_server}
\alias{module_file_selector_server}
\title{File selector server}
\usage{
module_file_selector_server(
  input,
  output,
  session,
  get_variable,
  get_iso_files
)
}
\arguments{
\item{get_variable}{reactive function returning the name of the variable}

\item{get_iso_files}{reactive function returning the actual iso files}
}
\description{
File selector server
}
