% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_selector.R
\name{module_file_selector_ui}
\alias{module_file_selector_ui}
\title{File selector UI}
\usage{
module_file_selector_ui(id, width = 12, file_list_height = "200px")
}
\arguments{
\item{id}{the module id}

\item{width}{box width}

\item{file_list_height}{height of the file list}
}
\description{
File selector UI
}
\seealso{
Other datasets module functions: 
\code{\link{problems_server}()}
}
\concept{datasets module functions}
