% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_info.R
\name{module_info_server}
\alias{module_info_server}
\title{info server}
\usage{
module_info_server(
  input,
  output,
  session,
  settings,
  get_variables,
  get_settings,
  reset_settings
)
}
\arguments{
\item{get_variables}{reactive function returning the variables}

\item{get_settings}{reactive function returning the settings}

\item{reset_settings}{function to trigger settings reset}
}
\description{
info server
}
