% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_navbar.R
\name{module_navbar_server}
\alias{module_navbar_server}
\title{Variable Selection Server}
\usage{
module_navbar_server(input, output, session, selected_variable = NULL)
}
\arguments{
\item{selected_variable}{the default variable to select (if any)}
}
\description{
Module to select environment variable
}
