% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_navbar.R
\name{module_navbar_ui}
\alias{module_navbar_ui}
\title{module variable selection navbar}
\usage{
module_navbar_ui(id)
}
\arguments{
\item{id}{the module id}
}
\description{
module variable selection navbar
}
