% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings.R
\name{module_settings_server}
\alias{module_settings_server}
\title{stores the settings}
\usage{
module_settings_server(
  input,
  output,
  session,
  settings,
  initial_settings = list(),
  store_in_global_env = FALSE
)
}
\arguments{
\item{initial_settings}{initial settings}

\item{store_in_global_env}{whether to load/save settings from/in global environment}
}
\description{
stores the settings
}
