% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_cf.R
\name{plot_cf_server}
\alias{plot_cf_server}
\title{cf plot}
\usage{
plot_cf_server(input, output, session, get_variable, get_iso_files, is_visible)
}
\arguments{
\item{get_variable}{get variable name}
}
\description{
cf plot
}
