% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_di.R
\name{plot_di_server}
\alias{plot_di_server}
\title{Dual inlet plot}
\usage{
plot_di_server(input, output, session, get_variable, get_iso_files, is_visible)
}
\arguments{
\item{get_variable}{get variable name}
}
\description{
Dual inlet plot
}
