% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_scan.R
\name{plot_scan_server}
\alias{plot_scan_server}
\title{scan plot}
\usage{
plot_scan_server(
  input,
  output,
  session,
  settings,
  get_variable,
  get_iso_files,
  is_visible
)
}
\arguments{
\item{get_variable}{get variable name}
}
\description{
scan plot
}
