% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot.R
\name{plot_server}
\alias{plot_server}
\title{generic plot server}
\usage{
plot_server(
  input,
  output,
  session,
  settings,
  get_variable,
  generate_plot,
  reset_trigger = reactive({ })
)
}
\arguments{
\item{get_variable}{get variable name}

\item{generate_plot}{function to create the plot}

\item{reset_trigger}{reactive function to trigger a reset of the plot options (plot height, legend and font size)}
}
\description{
generic plot server
}
