% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_download.R, R/module_plot.R
\name{plot_ui}
\alias{plot_ui}
\title{data download ui}
\usage{
plot_ui(
  id,
  min_height = "500px;",
  action_widths = c(3, 6, 3),
  left_actions = list(),
  center_actions = list(),
  right_actions = list(),
  click = FALSE,
  dblclick = FALSE,
  brush_direction = NULL
)

plot_ui(
  id,
  min_height = "500px;",
  action_widths = c(3, 6, 3),
  left_actions = list(),
  center_actions = list(),
  right_actions = list(),
  click = FALSE,
  dblclick = FALSE,
  brush_direction = NULL
)
}
\arguments{
\item{brush_direction}{if NULL no brush, if "x", "y", or "xy", creates a brush}
}
\description{
data download ui

plot UI
}
