% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ggplot_download.R
\name{plotDownloadServer}
\alias{plotDownloadServer}
\title{Plot Download Server}
\usage{
plotDownloadServer(input, output, session, plot_func, filename_func)
}
\arguments{
\item{plot_func}{reactive function generating the plot}

\item{filename_func}{reactive function returning the default plot name}
}
\description{
Plot Download Server
}
