% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_selector.R
\name{problems_button}
\alias{problems_button}
\title{Dataset problems button}
\usage{
problems_button(id, tooltip = "Show problems.")
}
\arguments{
\item{tooltip}{what tooltip to display}
}
\description{
Dataset problems button
}
