% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_selector.R
\name{problems_server}
\alias{problems_server}
\title{Problems Server
Stand-alone for showing a dataset's problems.}
\usage{
problems_server(input, output, session, get_variable, get_iso_files)
}
\arguments{
\item{get_variable}{reactive function returning the name of the variable}

\item{get_iso_files}{reactive function returning the actual iso files}
}
\description{
Problems Server
Stand-alone for showing a dataset's problems.
}
\seealso{
Other datasets module functions: 
\code{\link{module_file_selector_ui}()}
}
\concept{datasets module functions}
