% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_datasets.R
\name{problemsButton}
\alias{problemsButton}
\title{Dataset problems button}
\usage{
problemsButton(id, tooltip = "Show problems.")
}
\arguments{
\item{id}{the file selector id}

\item{tooltip}{what tooltip to display}

\item{start_disabled}{whether to start out with the button disabled}
}
\description{
Dataset problems button
}
\seealso{
Other datasets module functions: \code{\link{datasetsServer}},
  \code{\link{datasetsUI}}, \code{\link{problemsServer}}
}
