% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_datasets.R
\name{problemsServer}
\alias{problemsServer}
\title{Problems Server
Stand-alone for showing a dataset's problems.}
\usage{
problemsServer(input, output, session, dataset, dataset_path)
}
\arguments{
\item{dataset}{reactive function returning the isofiles list}

\item{dataset_path}{reactive function returning the dataset's path}
}
\description{
Problems Server
Stand-alone for showing a dataset's problems.
}
\seealso{
Other datasets module functions: \code{\link{datasetsServer}},
  \code{\link{datasetsUI}}, \code{\link{problemsButton}}
}
