% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{refresh_directory_tree}
\alias{refresh_directory_tree}
\title{Refresh directory tree}
\usage{
refresh_directory_tree(storage, verbose = FALSE)
}
\value{
storage object with directory_tree updated
}
\description{
Refresh directory tree
}
\seealso{
Other data storage functions: \code{\link{gdrive_storage}},
  \code{\link{get_directory_tree}},
  \code{\link{get_recent_files}},
  \code{\link{init_storage}}, \code{\link{local_storage}},
  \code{\link{refresh_recent_files}}
}
