% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{refresh_recent_files}
\alias{refresh_recent_files}
\title{Refresh recent files}
\usage{
refresh_recent_files(storage, file_pattern, n_max, verbose = FALSE)
}
\arguments{
\item{storage}{local storage object}

\item{file_pattern}{what types of files to return}

\item{n_max}{how many recent files to fetch}

\item{verbose}{whether to provide information messages or not}
}
\value{
storage object with recent_files updated
}
\description{
Refresh recent files
}
\seealso{
Other data storage functions: \code{\link{gdrive_storage}},
  \code{\link{get_directory_tree}},
  \code{\link{get_recent_files}},
  \code{\link{init_storage}}, \code{\link{local_storage}},
  \code{\link{refresh_directory_tree}}
}
