% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{run}
\alias{run}
\title{Start the user interface}
\usage{
run(data_dir = ".", ..., launch = TRUE)
}
\arguments{
\item{data_dir}{the directory for local data files}

\item{...}{passed on to the \code{\link[shiny]{runApp}} call (only if \code{launch = TRUE}), can include server-specific parameters such as host or port}

\item{launch}{whether to launch the app (TRUE) or return a shiny app object (FALSE) that then can be launched via \code{\link[shiny]{runApp}}}
}
\description{
Start the user interface
}
