% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{select_current_folder}
\alias{select_current_folder}
\title{Select a current folder in the directory tree}
\usage{
select_current_folder(storage, folder)
}
\arguments{
\item{folder}{which folder to select}
}
\description{
Select a current folder in the directory tree
}
