% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_selector_table.R
\name{selector_table_buttons_ui}
\alias{selector_table_buttons_ui}
\title{Selector table buttons}
\usage{
selector_table_buttons_ui(id)
}
\description{
Selector table buttons
}
\seealso{
Other selector table module functions: 
\code{\link{selector_table_server}()},
\code{\link{selector_table_ui}()}
}
\concept{selector table module functions}
