% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_selector_table.R
\name{selector_table_ui}
\alias{selector_table_ui}
\title{Selector table UI}
\usage{
selector_table_ui(id)
}
\description{
Selector table UI
}
\seealso{
Other selector table module functions: 
\code{\link{selector_table_buttons_ui}()},
\code{\link{selector_table_server}()}
}
\concept{selector table module functions}
