% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_selector_table.R
\name{selectorTableButtons}
\alias{selectorTableButtons}
\title{Selector table buttons}
\usage{
selectorTableButtons(id)
}
\arguments{
\item{id}{the module id}
}
\description{
Selector table buttons
}
\seealso{
Other selector table module functions: \code{\link{selectorTableServer}},
  \code{\link{selectorTableUI}}
}
