% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_selector_table.R
\name{selectorTableServer}
\alias{selectorTableServer}
\title{Selector table server}
\usage{
selectorTableServer(input, output, session, id_column, col_headers,
  hot_mods = NULL)
}
\arguments{
\item{id_column}{name of the id column that should be used for reporting selection}

\item{col_headers}{name of the column headers}

\item{hot_mods}{hands on table modifiers}
}
\description{
This generates an rhandson table for selecting items to include in downstream operations.
}
\seealso{
Other selector table module functions: \code{\link{selectorTableButtons}},
  \code{\link{selectorTableUI}}
}
