% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_selector_table.R
\name{selectorTableUI}
\alias{selectorTableUI}
\title{Selector table UI}
\usage{
selectorTableUI(id, height = "200px")
}
\arguments{
\item{id}{the module id}
}
\description{
Selector table UI
}
\seealso{
Other selector table module functions: \code{\link{selectorTableButtons}},
  \code{\link{selectorTableServer}}
}
