% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_trace_selector.R
\name{trace_selector_server}
\alias{trace_selector_server}
\title{select data trace}
\usage{
trace_selector_server(input, output, session, get_variable, get_iso_files)
}
\description{
select data trace
}
