% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_trace_selector.R
\name{trace_selector_table_buttons_ui}
\alias{trace_selector_table_buttons_ui}
\title{trace selector buttons ui}
\usage{
trace_selector_table_buttons_ui(id)
}
\description{
trace selector buttons ui
}
