% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_vendor_data_table.R
\name{vendorDataTableSelectorUI}
\alias{vendorDataTableSelectorUI}
\title{File Info Selector UI}
\usage{
vendorDataTableSelectorUI(id, width = 4, selector_height = "200px")
}
\arguments{
\item{id}{the module id}

\item{width}{box width}

\item{selector_height}{file selector height}
}
\description{
File Info Selector UI
}
\seealso{
Other vendor data table module functions: \code{\link{vendorDataTableServer}},
  \code{\link{vendorDataTableTableUI}}
}
