% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_vendor_data_table.R
\name{vendorDataTableServer}
\alias{vendorDataTableServer}
\title{Vendor Data Table Server}
\usage{
vendorDataTableServer(input, output, session, isofiles, visible = NULL)
}
\arguments{
\item{isofiles}{reactive function returning the currently loaded isofiles}

\item{visible}{reactive function determining visibility of the auxiliary boxes}
}
\description{
Vendor Data Table Server
}
\seealso{
Other vendor data table module functions: \code{\link{vendorDataTableSelectorUI}},
  \code{\link{vendorDataTableTableUI}}
}
