% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_vendor_data_table.R
\name{vendorDataTableTableUI}
\alias{vendorDataTableTableUI}
\title{File Info Table UI}
\usage{
vendorDataTableTableUI(id)
}
\arguments{
\item{id}{the module id}
}
\description{
File Info Table UI
}
\seealso{
Other vendor data table module functions: \code{\link{vendorDataTableSelectorUI}},
  \code{\link{vendorDataTableServer}}
}
