% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{viewer_server}
\alias{viewer_server}
\title{Isoviewer App Server}
\usage{
viewer_server(iso_objects = list(), local = FALSE)
}
\arguments{
\item{iso_objects}{named list of iso file objects to make available in the viewer}

\item{local}{whether running locally or as a server application (e.g. on shinyapps.io)}
}
\description{
Generates the server part of the isoviewer app
}
