% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{viewer_ui}
\alias{viewer_ui}
\title{Isoviewer App UI}
\usage{
viewer_ui(sidebar_width = 170, start_menu = "welcome", theme = "united")
}
\arguments{
\item{sidebar_width}{the width of the sidebar}

\item{deault_menu}{default selected menu}
}
\description{
Generates the user interface part of the isoviewer app
}
