% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Isoviewer App Server}
\usage{
app_server(data_dir, allow_data_upload, allow_folder_creation, store_data)
}
\arguments{
\item{data_dir}{the directory for local data files}

\item{allow_data_upload}{whether to allow uploading of data}

\item{allow_folder_creation}{whether to allow creation of folders on the server}

\item{store_data}{whether data files (including .rda exports) are stored permanently (TRUE) or just temporarily (FALSE)}
}
\description{
Generates the server part of the isoviewer app
}
