% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_upload.R
\name{dataUploadServer}
\alias{dataUploadServer}
\title{Data Upload Server}
\usage{
dataUploadServer(input, output, session, folder, show_folder = NULL)
}
\arguments{
\item{folder}{should be a reactive function that returns the upload folder}

\item{show_folder}{reactive function that shows the view folder}
}
\description{
Data Upload Server
}
\seealso{
Other data upload module functions: \code{\link{dataUploadUI}}
}
