% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{iso_start_viewer}
\alias{iso_start_viewer}
\title{Start the isoviewer graphical user interface}
\usage{
iso_start_viewer(
  iso_objects = iso_find_objects(),
  local = TRUE,
  log = !local,
  runApp_params = list()
)
}
\arguments{
\item{iso_objects}{named list of iso file objects to make available in the viewer}

\item{local}{whether running locally or as a server application (e.g. on shinyapps.io)}

\item{log}{whether to show log info messages or not (by default only if running as server application, i.e. non-local)}

\item{runApp_params}{passed on to the \code{\link[shiny]{runApp}} call. Common parameters to specify include \code{port} and \code{launch.browser}. Only relevant if \code{local = TRUE}.}
}
\description{
Searches for all iso file objects in the workspace and provides a graphical interface to exlore the data stored in them.
}
