% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{run}
\alias{run}
\title{Start the user interface}
\usage{
run(data_dir = ".", allow_data_upload = FALSE,
  allow_folder_creation = FALSE, store_data = TRUE, ..., launch = TRUE)
}
\arguments{
\item{data_dir}{the directory for local data files}

\item{allow_data_upload}{whether to allow uploading of data}

\item{allow_folder_creation}{whether to allow creation of folders on the server}

\item{store_data}{whether data files (including .rda exports) are stored permanently (TRUE) or just temporarily (FALSE)}

\item{...}{passed on to the \code{\link[shiny]{runApp}} call (only if \code{launch = TRUE}), can include server-specific parameters such as host or port}

\item{launch}{whether to launch the app (TRUE) or return a shiny app object (FALSE) that then can be launched via \code{\link[shiny]{runApp}}}
}
\description{
Start the user interface
}
