% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run}
\alias{run}
\title{Start the user interface}
\usage{
run(data_dir = ".", ...)
}
\arguments{
\item{data_dir}{the directory for local data files}

\item{...}{passed on to the \code{\link[shiny]{runApp}} call, can include server-specific parameters such as host or port}
}
\description{
Start the user interface
}
